Added "include" mode
Now it's possible to choose between defining a list of buttons to
exclude from the widget or to include. When you use "include" the widget
will only trigger if the user navigates away from the page with one of
the selected buttons, making it less likely to trigger the widget
unintentionally.